A* search algorythm (disabled, not working properly)

+++ src/com/dtrajko/java/game/Game.java
@@ -144,9 +144,9 @@ public class Game extends Canvas implements Runnable {
 		}
 
 		screen.clear();
-		int xScroll = player.getX() - screen.width / 2;
-		int yScroll = player.getY() - screen.height / 2;
-		level.render(xScroll, yScroll, screen);
+		double xScroll = player.getX() - screen.width / 2;
+		double yScroll = player.getY() - screen.height / 2;
+		level.render((int) xScroll, (int) yScroll, screen);
 		// player.render(screen);
 		// mob.render(screen);
 		// screen.render(x, y);
