+++ src/com/dtrajko/java/game/entity/Entity.java
@@ -11,7 +11,7 @@ import com.dtrajko.java.game.level.Level;
  */
 public abstract class Entity {
 
-	protected int x, y;
+	protected double x, y;
 	protected Sprite sprite;
 	private boolean removed = false;
 	protected Level level;
@@ -36,15 +36,15 @@ public abstract class Entity {
 
 	public void render(Screen screen) {
 		if (sprite != null) {
-			screen.renderSprite(x, y, sprite, true);
+			screen.renderSprite((int) x, (int) y, sprite, true);
 		}
 	}
 	
-	public int getX() {
+	public double getX() {
 		return this.x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return this.y;
 	}
