+++ src/com/dtrajko/java/game/entity/projectile/Projectile.java
@@ -5,14 +5,14 @@ import com.dtrajko.java.game.graphics.Sprite;
 
 public abstract class Projectile extends Entity {
 
-	protected final int xOrigin, yOrigin;
+	protected final double xOrigin, yOrigin;
 	protected double angle;
 	protected Sprite sprite;
 	protected double x, y;
 	protected double nx, ny;
 	protected double speed, range, damage;
 
-	public Projectile(int x, int y, double dir) {
+	public Projectile(double x, double y, double dir) {
 		xOrigin = x;
 		yOrigin = y;
 		angle = dir;
